+++ Editor/PropertyDrawers/AUEMethodParameterInfoPropertyDrawer.cs
@@ -10,12 +10,14 @@ namespace AUE
     [CustomPropertyDrawer(typeof(AUEMethodParameterInfo))]
     public class AUEMethodParameterInfoPropertyDrawer : PropertyDrawer
     {
+        private const int ModeWidth = 150;
+        private const int LabelModeSpace = 10;
+
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
             float height = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
             if (property.isExpanded)
             {
-                height += EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
                 var customMethodSP = property.FindPropertyRelative(AUEUtils.CustomArgumentSPName);
                 height += EditorGUI.GetPropertyHeight(customMethodSP, customMethodSP.isExpanded);
             }
@@ -27,41 +29,63 @@ namespace AUE
             Rect lineRect = position;
             lineRect.height = EditorGUIUtility.singleLineHeight;
 
-            property.isExpanded = EditorGUI.Foldout(lineRect, property.isExpanded, label);
-            lineRect.y += EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+            Rect labelRect = new Rect(lineRect.x, lineRect.y, lineRect.width - ModeWidth - LabelModeSpace, lineRect.height);
+            Rect modeRect = new Rect(lineRect.xMax - ModeWidth, lineRect.y, ModeWidth, lineRect.height);
+            property.isExpanded = EditorGUI.Foldout(labelRect, property.isExpanded, label);
+
+            var customArgumentSP = property.FindPropertyRelative(AUEUtils.CustomArgumentSPName);
+            var modeSP = property.FindPropertyRelative(AUEUtils.ModeSPName);
+            InitializeCustomArgumentIFN(customArgumentSP, (EMode)modeSP.enumValueIndex);
+
+            EditorGUI.BeginChangeCheck();
+            {
+                using (new PropertyDrawerHelper.IndentedLevelResetScope())
+                {
+                    EditorGUI.PropertyField(modeRect, modeSP, GUIContent.none);
+                }
+            }
+            if (EditorGUI.EndChangeCheck())
+            {
+                UpdateCustomArgumentType(property, customArgumentSP, (EMode)modeSP.enumValueIndex);
+                property.serializedObject.ApplyModifiedProperties();
+                property.serializedObject.Update();
+            }
 
             if (property.isExpanded)
             {
+                lineRect.y += EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
                 ++EditorGUI.indentLevel;
-
-                var customArgumentSP = property.FindPropertyRelative(AUEUtils.CustomArgumentSPName);
-                var modeSP = property.FindPropertyRelative(AUEUtils.ModeSPName);
-                InitializeCustomArgumentIFN(customArgumentSP, (EMode)modeSP.enumValueIndex);
-
-                EditorGUI.BeginChangeCheck();
                 {
-                    EditorGUI.PropertyField(lineRect, modeSP);
-                    lineRect.y += EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+                    Rect propRect = new Rect(lineRect.x, lineRect.y, lineRect.width, EditorGUI.GetPropertyHeight(customArgumentSP, includeChildren: customArgumentSP.isExpanded));
+                    var propertyDrawer = customArgumentSP.GetPropertyDrawer();
+                    propertyDrawer.OnGUI(propRect, customArgumentSP, new GUIContent(customArgumentSP.displayName));
                 }
-                if (EditorGUI.EndChangeCheck())
-                {
-                    UpdateCustomArgumentType(property, customArgumentSP, (EMode)modeSP.enumValueIndex);
-                }
-
-                Rect propRect = new Rect(lineRect.x, lineRect.y, lineRect.width, EditorGUI.GetPropertyHeight(customArgumentSP, includeChildren: customArgumentSP.isExpanded));
-                EditorGUI.PropertyField(propRect, customArgumentSP, includeChildren: customArgumentSP.isExpanded);
-
                 --EditorGUI.indentLevel;
             }
         }
 
-        private void UpdateCustomArgumentType(SerializedProperty parameterInfoSP, SerializedProperty customArgumentSP, EMode mode)
+        public static void InitializeCustomArgument(SerializedProperty parameterInfoSP)
+        {
+            var customArgumentSP = parameterInfoSP.FindPropertyRelative(AUEUtils.CustomArgumentSPName);
+            var modeSP = parameterInfoSP.FindPropertyRelative(AUEUtils.ModeSPName);
+            UpdateCustomArgumentType(parameterInfoSP, customArgumentSP, (EMode)modeSP.intValue);
+        }
+
+        private static void UpdateCustomArgumentType(SerializedProperty parameterInfoSP, SerializedProperty customArgumentSP, EMode mode)
         {
             TryDeleteMethodFromDatabase(parameterInfoSP, customArgumentSP);
             InitializeCustomArgument(customArgumentSP, mode);
         }
 
-        private void InitializeCustomArgument(SerializedProperty customArgumentSP, EMode mode)
+        private static void InitializeCustomArgumentIFN(SerializedProperty customArgumentSP, EMode mode)
+        {
+            if (string.IsNullOrEmpty(customArgumentSP.managedReferenceFullTypename))
+            {
+                InitializeCustomArgument(customArgumentSP, mode);
+            }
+        }
+
+        private static void InitializeCustomArgument(SerializedProperty customArgumentSP, EMode mode)
         {
             object refValue = null;
             switch (mode)
@@ -77,13 +101,17 @@ namespace AUE
                     break;
             }
             customArgumentSP.managedReferenceValue = refValue;
-        }
-
-        private void InitializeCustomArgumentIFN(SerializedProperty customArgumentSP, EMode mode)
-        {
-            if (string.IsNullOrEmpty(customArgumentSP.managedReferenceFullTypename))
+            switch (mode)
             {
-                InitializeCustomArgument(customArgumentSP, mode);
+                case EMode.Dynamic:
+                    AUECADynamicPropertyDrawer.Initialize(customArgumentSP);
+                    break;
+                case EMode.Constant:
+                    AUECAConstantPropertyDrawer.Initialize(customArgumentSP);
+                    break;
+                case EMode.Method:
+                    AUECAMethodReferencePropertyDrawer.Initialize(customArgumentSP);
+                    break;
             }
         }