+++ Editor/PropertyDrawers/AUEMethodPropertyDrawer.cs
@@ -13,17 +13,32 @@ namespace AUE
             public InvokeInfo NewInvokeInfo = null;
         }
 
+        private const int CallStateModeWidth = 150;
+        private const int TargetCallStateSpace = 10;
+
         private const string MethodNameSPName = "_methodName";
         private const string TargetSPName = "_target";
         private const string ParameterInfosSPName = "_parameterInfos";
         private const string ParameterTypeSPName = "_parameterType";
         private const string CallStateSPName = "_callState";
+        private const string StaticTypeSPName = "_staticType";
 
         private Dictionary<string, PropertyMetaData> _metaData = new Dictionary<string, PropertyMetaData>();
 
+        private static GUIStyle MethodButtonStyle;
+
+        private void InitializePropertyDrawerIFN()
+        {
+            if (MethodButtonStyle == null)
+            {
+                MethodButtonStyle = new GUIStyle(GUI.skin.button);
+                MethodButtonStyle.alignment = TextAnchor.MiddleLeft;
+            }
+        }
+
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
-            float height = (EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing) * 2;
+            float height = (EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing) * 1;
 
             var targetSP = property.FindPropertyRelative(TargetSPName);
             if (targetSP.objectReferenceValue != null)
@@ -52,16 +67,30 @@ namespace AUE
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
+            InitializePropertyDrawerIFN();
+
             Rect lineRect = position;
             lineRect.height = EditorGUIUtility.singleLineHeight;
 
-            EditorGUI.PropertyField(lineRect, property.FindPropertyRelative(CallStateSPName));
-            lineRect.y += lineRect.height + EditorGUIUtility.standardVerticalSpacing;
+            var callStateSP = property.FindPropertyRelative(CallStateSPName);
+
+            Rect targetRect = lineRect;
+            if (callStateSP != null)
+            {
+                targetRect.width = targetRect.width - CallStateModeWidth - TargetCallStateSpace;
+
+                Rect callStateRect = new Rect(targetRect.xMax + TargetCallStateSpace, targetRect.y, CallStateModeWidth, lineRect.height);
+                using (new PropertyDrawerHelper.IndentedLevelResetScope())
+                {
+                    EditorGUI.PropertyField(callStateRect, callStateSP, GUIContent.none);
+                }
+            }
 
             EditorGUI.BeginChangeCheck();
             var targetSP = property.FindPropertyRelative(TargetSPName);
-            EditorGUI.PropertyField(lineRect, targetSP);
             lineRect.y += lineRect.height + EditorGUIUtility.standardVerticalSpacing;
+            EditorGUI.PropertyField(targetRect, targetSP, GUIContent.none);
+
             if (EditorGUI.EndChangeCheck())
             {
                 UpdateMethodName(property, null);
@@ -88,8 +117,12 @@ namespace AUE
             var methodNameSP = property.FindPropertyRelative(MethodNameSPName);
             if (AUEUtils.LoadMethodInfoFromAUEMethod(property, out TargetInvokeInfo[] invokeInfos, out InvokeInfo selectedInvoke))
             {
-                string method = (selectedInvoke != null ? selectedInvoke.MethodMeta.DisplayName : "<None>");
-                if (GUI.Button(position, new GUIContent(method)))
+                string method = selectedInvoke != null ? 
+                    MethodPreviewBuilder.GenerateMethodPreview(property, selectedInvoke.MethodMeta.MethodInfo) :
+                    "<None>";
+
+                Rect indentedPosition = EditorGUI.IndentedRect(position);
+                if (GUI.Button(indentedPosition, new GUIContent(method), MethodButtonStyle))
                 {
                     var dropdown = new MethodSearchDropdown(property, invokeInfos,
                         (prop, newInvokeInfo) =>
@@ -99,7 +132,7 @@ namespace AUE
                             var metaData = CreateTempMetaData(prop);
                             metaData.NewInvokeInfo = newInvokeInfo;
                         });
-                    dropdown.Show(position);
+                    dropdown.Show(indentedPosition);
                 }
 
                 var metaData = TryGetTempMetaData(property);
@@ -114,10 +147,12 @@ namespace AUE
                             var methodMetaData = newInvokeInfo.MethodMeta;
                             UpdateMethodName(property, methodMetaData);
                             UpdateTarget(property, newInvokeInfo.Target);
+                            UpdateStaticType(property, newInvokeInfo.Target);
                             UpdateParameterInfos(property, methodMetaData.MethodInfo);
                         }
                         else
                         {
+                            UpdateStaticType(property, null);
                             UpdateMethodName(property, null);
                             UpdateParameterInfos(property, null);
                         }
@@ -129,6 +164,20 @@ namespace AUE
                 }
 
                 position.y += position.height + EditorGUIUtility.standardVerticalSpacing;
+            }
+        }
+
+        private void UpdateStaticType(SerializedProperty property, Object target)
+        {
+            var staticTypeSP = property.FindPropertyRelative(StaticTypeSPName);
+            if (target is MonoScript)
+            {
+                var targetType = AUEUtils.GetTargetType(target);
+                SerializableTypeHelper.SetType(staticTypeSP, targetType);
+            }
+            else
+            {
+                SerializableTypeHelper.ClearType(staticTypeSP);
             }
         }
 
@@ -166,10 +215,10 @@ namespace AUE
                         float height = EditorGUI.GetPropertyHeight(parameterInfoSP);
                         Rect propRect = new Rect(position.x, position.y, position.width, height);
                         GUI.Box(propRect, GUIContent.none);
-                        EditorGUI.PropertyField(propRect, 
-                            parameterInfoSP, 
-                            new GUIContent($"{AUEUtils.MakeHumanDisplayType(parameterInfos[i].ParameterType)} {parameterInfos[i].Name}"),
-                            parameterInfosSP.isExpanded);
+
+                        var propertyDrawer = parameterInfoSP.GetPropertyDrawer();
+                        propertyDrawer.OnGUI(propRect, parameterInfoSP, new GUIContent($"{AUEUtils.MakeHumanDisplayType(parameterInfos[i].ParameterType)} {parameterInfos[i].Name}"));
+
                         position.y += height + EditorGUIUtility.standardVerticalSpacing;
                     }
                     --EditorGUI.indentLevel;
@@ -203,13 +252,15 @@ namespace AUE
         private static void InitializeParameterInfo(SerializedProperty parameterInfoSP, System.Type parameterType)
         {
             var parameterTypeSP = parameterInfoSP.FindPropertyRelative(ParameterTypeSPName);
-            SerializableTypeHelper.SetTypeName(parameterTypeSP, parameterType);
+            SerializableTypeHelper.SetType(parameterTypeSP, parameterType);
 
             var modeSP = parameterInfoSP.FindPropertyRelative(AUEUtils.ModeSPName);
             modeSP.enumValueIndex = (int)AUEMethodParameterInfo.EMode.Constant;
 
             var customArgumentSP = parameterInfoSP.FindPropertyRelative(AUEUtils.CustomArgumentSPName);
             customArgumentSP.managedReferenceValue = null;
+
+            AUEMethodParameterInfoPropertyDrawer.InitializeCustomArgument(parameterInfoSP);
         }
 
         private void SetMethodDirty(SerializedProperty property)