+++ Runtime/Core/AUEMethod.cs
@@ -5,12 +5,17 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using UnityEngine.Assertions;
+using UnityEngine.Events;
 
 namespace AUE
 {
     [Serializable]
     public class AUEMethod : AUESimpleMethod, IMethodDatabaseOwner
     {
+        [SerializeField]
+        protected UnityEventCallState _callState = UnityEventCallState.RuntimeOnly;
+        public UnityEventCallState CallState { get => _callState; set => _callState = value; }
+
         [SerializeField]
         private List<SerializableType> _argumentTypes = new List<SerializableType>();
         public List<SerializableType> ArgumentTypes
@@ -25,6 +30,10 @@ namespace AUE
 
         internal object Invoke(params object[] args)
         {
+            if (!CanBeExecuted())
+            {
+                return null;
+            }
             return base.Invoke(this, args);
         }
 
@@ -51,6 +60,20 @@ namespace AUE
             {
                 _parameterInfos[i] = new AUEMethodParameterInfo(desc.Parameters[i], parameters[i].ParameterType);
             }
+        }
+
+        private bool CanBeExecuted()
+        {
+            if (_callState == UnityEventCallState.Off)
+            {
+                return false;
+            }
+            else if (_callState == UnityEventCallState.RuntimeOnly)
+            {
+                return UnityEngine.Application.isPlaying;
+            }
+
+            return true;
         }
 
         private static Type[] GetParameterTypes(AUEParameterDescriptor[] parameters)